actions: Move part into zerver.lib.streams.
(cherry picked from commit a29f1b39da1e9b4b283e8000645ee877548f30e7)
