+++ tools/linter_lib/custom_check.py
@@ -379,7 +379,6 @@ python_rules = RuleList(
                 # This one in check_message is kinda terrible, since it's
                 # how most instances are written, but better to exclude something than nothing
                 ("zerver/lib/actions.py", "stream = get_stream(stream_name, realm)"),
-                ("zerver/lib/actions.py", 'return get_stream("signups", realm)'),
             },
             "description": "Please use access_stream_by_*() to fetch Stream objects",
         },
