+++ zerver/lib/events.py
@@ -13,12 +13,7 @@ from zerver.actions.default_streams import (
     get_default_streams_for_realm,
     streams_to_dicts_sorted,
 )
-from zerver.lib.actions import (
-    do_get_streams,
-    gather_subscriptions_helper,
-    get_owned_bot_dicts,
-    get_web_public_streams,
-)
+from zerver.lib.actions import gather_subscriptions_helper, get_owned_bot_dicts
 from zerver.lib.alert_words import user_alert_words
 from zerver.lib.avatar import avatar_url
 from zerver.lib.bot_config import load_bot_config_template
@@ -46,6 +41,7 @@ from zerver.lib.realm_logo import get_realm_logo_source, get_realm_logo_url
 from zerver.lib.soft_deactivation import reactivate_user_if_soft_deactivated
 from zerver.lib.sounds import get_available_notification_sounds
 from zerver.lib.stream_subscription import handle_stream_notifications_compatibility
+from zerver.lib.streams import do_get_streams, get_web_public_streams
 from zerver.lib.subscription_info import get_web_public_subs
 from zerver.lib.timestamp import datetime_to_timestamp
 from zerver.lib.timezone import canonicalize_timezone
