+++ zerver/management/commands/create_default_stream_groups.py
@@ -1,8 +1,8 @@
 from argparse import ArgumentParser
 from typing import Any
 
-from zerver.lib.actions import ensure_stream
 from zerver.lib.management import ZulipBaseCommand
+from zerver.lib.streams import ensure_stream
 from zerver.models import DefaultStreamGroup
