+++ zerver/tests/test_email_mirror.py
@@ -11,12 +11,7 @@ import orjson
 from django.conf import settings
 from django.http import HttpResponse
 
-from zerver.lib.actions import (
-    do_change_stream_post_policy,
-    do_deactivate_realm,
-    do_deactivate_user,
-    ensure_stream,
-)
+from zerver.lib.actions import do_change_stream_post_policy, do_deactivate_realm, do_deactivate_user
 from zerver.lib.email_mirror import (
     create_missed_message_address,
     filter_footer,
@@ -37,6 +32,7 @@ from zerver.lib.email_mirror_helpers import (
 )
 from zerver.lib.email_notifications import convert_html_to_markdown
 from zerver.lib.send_email import FromAddress
+from zerver.lib.streams import ensure_stream
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import mock_queue_publish, most_recent_message, most_recent_usermessage
 from zerver.models import (
