+++ zerver/views/streams.py
@@ -38,7 +38,6 @@ from zerver.lib.actions import (
     do_change_subscription_property,
     do_deactivate_stream,
     do_delete_messages,
-    do_get_streams,
     do_rename_stream,
     do_send_messages,
     gather_subscriptions,
@@ -64,6 +63,7 @@ from zerver.lib.streams import (
     access_stream_for_delete_or_update,
     access_web_public_stream,
     check_stream_name_available,
+    do_get_streams,
     filter_stream_authorization,
     get_stream_permission_policy_name,
     list_to_streams,
